dormitory-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.1
Werkzeug==3.0.1
python-dotenv==1.0.1
Flask-Mail==0.9.1
apscheduler==3.10.4
gunicorn==22.0.0
Pillow==10.3.0
pydantic==2.5.2
flask-limiter==3.5.0
flask-caching==2.0.2
celery==5.3.4
redis==5.0.1
flask-swagger-ui==4.11.1
sentry-sdk==1.39.1
requests==2.31.0
email-validator==2.2.0
unidecode==1.3.8
bleach==6.1.0